pyproject.toml
[project]
name = "anymusicbot"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "wheel>=0.43.0,<0.44",
    "setuptools>=72.1.0,<73",
    "aiogram>=3.1.1,<4",
    "rich>=13.6.0,<14",
    "shazamio",
    "sqlitedict>=2.1.0,<3",
    "spotipy>=2.23.0,<3",
    "attrs>=23.1.0,<24",
    "ytmusicapi>=1.8.0,<2",
    "pydub>=0.25.1,<0.26",
    "aiohttp>=3.8.6,<4",
    "nest-asyncio>=1.5.8,<2",
    "icecream>=2.1.3,<3",
    "m3u8>=5.1.0,<6",
    "cryptography>=43.0.0,<44",
    "pytubefix>=8.2.0,<9",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
shazamio = { path = "lib/ShazamIO" }
